link meals, extras and addons

+++ Api_restaurant/routes/diningTable.php
@@ -4,13 +4,13 @@ use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Route;
 use App\Http\Controllers\DiningTableController;
 
-Route::get('dining-tables', [DiningTableController::class, 'getAllDiningTables']);
 Route::get('dining-tables/{id}', [DiningTableController::class, 'getDiningTableById']);
-
-Route::post('dining-tables', [DiningTableController::class, 'addNewDiningTable']);
-Route::put('dining-tables/{id}', [DiningTableController::class, 'updateDiningTable']);
-Route::delete('dining-tables/{id}', [DiningTableController::class, 'deleteDiningTable']);
 
 Route::group(['middleware'=>'checkAdminToken'],function(){
 
+    Route::get('admin/dining-tables', [DiningTableController::class, 'getAllDiningTables']);
+    Route::post('admin/dining-tables', [DiningTableController::class, 'addNewDiningTable']);
+    Route::put('admin/dining-tables/{id}', [DiningTableController::class, 'updateDiningTable']);
+    Route::delete('admin/dining-tables/{id}', [DiningTableController::class, 'deleteDiningTable']);
+    
 });
